requirements.txt
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
langgraph>=0.1.0
google-search-results>=2.4.2
